+++ src/main/kotlin/com/yourssu/soongpt/message/storage/exception/MessageNotFoundException.kt
@@ -1,0 +1,5 @@
+package com.yourssu.soongpt.message.storage.exception
+
+import com.yourssu.soongpt.common.handler.NotFoundException
+
+class MessageNotFoundException : NotFoundException(message = "해당하는 메세지가 없습니다.")
